Factor out esm config
