+++ packages/modular-scripts/react-scripts/config/utils/esmUtils.js
@@ -1,6 +1,17 @@
 'use strict';
 
-const modules = require('../modules');
+const packageRegex =
+  /^(@[a-z0-9-~][a-z0-9-._~]*)?\/?([a-z0-9-~][a-z0-9-._~]*)\/?(.*)/;
+
+function parsePackageName(name) {
+  const parsedName = packageRegex.exec(name);
+  if (!parsedName) {
+    return;
+  }
+  const [_, scope, module, submodule] = parsedName;
+  const dependencyName = (scope ? `${scope}/` : '') + module;
+  return { dependencyName, scope, module, submodule };
+}
 
 function createExternalDependenciesMap({
   externalDependencies,
@@ -38,4 +49,5 @@ function createExternalDependenciesMap({
 
 module.exports = {
   createExternalDependenciesMap,
+  parsePackageName,
 };
